# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CDW Finance — API pública
  version: 1.0.0
  description: |
    API pública de **consulta** do CDW Finance, versionada sob `/api/v1`.

    Expõe os dados de gestão da empresa vinculada à chave: clientes, contratos,
    contatos, anexos, servidores e domínios. **Todas as operações são somente
    leitura** — não há escrita nesta versão.

    Os mesmos dados estão disponíveis por **MCP** (Model Context Protocol), para
    agentes de IA, em `POST /api/v1/mcp`. As permissões e o formato de cada
    entidade são idênticos nos dois caminhos. Instruções em `/mcp/docs`.

    ### Autenticação

    Toda requisição exige uma chave de API no cabeçalho:

    ```
    Authorization: Bearer cdwf_live_<prefixo>_<segredo>
    ```

    As chaves são geradas por empresa no painel, em **Empresas › Chaves de API**.
    O sistema guarda apenas o hash: a chave completa é exibida uma única vez, na
    geração. Cada chave carrega uma lista de escopos e pode ser revogada a
    qualquer momento.

    **O escopo por empresa é derivado da chave no servidor.** Nenhuma operação
    aceita `id_company` por query, corpo ou rota. Consultar um recurso de outra
    empresa devolve `404`, e não `403` — um `403` confirmaria que aquele
    registro existe.

    Use a chave sempre a partir do servidor: não há CORS, e expô-la no
    JavaScript de um navegador entregaria acesso a toda a base do tenant.

    ### Paginação

...
    detalhe não trazem `meta`.

    ### Limite de requisições

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /customers/{id}:
    get:
      tags: [Clientes]
      summary: Consulta um cliente por ID
      description: Requer o escopo `customers:read`. Cliente de outra empresa responde `404`.
      operationId: getCustomer
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Consulta realizada.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Cliente' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '403': { $ref: '#/components/responses/SemPermissao' }
        '404': { $ref: '#/components/responses/NaoEncontrado' }
        '422': { $ref: '#/components/responses/NaoProcessavel' }
        '429': { $ref: '#/components/responses/LimiteExcedido' }
...
  /contracts/{id}:
    get:
      tags: [Contratos]
      summary: Consulta a visão geral de um contrato
      description: |
        Requer o escopo `contracts:read`. Além dos dados do contrato, devolve
        os tipos de serviço e os **domínios contratados**, no campo `domains`.

        O bloco de domínios não é paginado e traz no máximo 200 linhas; para
        listas maiores, use `/contract-domains?contract_id=`.

        ### Títulos

        Com o escopo **`invoices:read`**, a resposta inclui também os dois
        blocos de título, que vêm de origens diferentes:

        - **`invoices`** — as faturas que o CDW Finance gerou, do banco local.
          No máximo 200, da mais recente para a mais antiga.
        - **`bomcontrole_invoices`** — o extrato do ERP Bom Controle,
          consultado **ao vivo**. Nada de financeiro do ERP é copiado para o
          banco, para não criar uma segunda verdade sobre faturas que mudam lá.

        O ERP só é consultado quando o contrato está **vinculado** a um
        contrato de venda de lá; sem vínculo, o bloco volta com
        `linked: false` e nenhuma chamada de rede acontece. Falha na consulta
        **não derruba a resposta**: o bloco sai com `available: false` e o
        motivo, e o resto do contrato é servido normalmente.

        Um contrato normalmente tem só uma das duas origens populadas —
        `billing.source` diz quem fatura, e é exclusivo. As duas cheias são o
        retrato de uma transição.

        **Sem o escopo `invoices:read` os dois blocos são omitidos** e a
        requisição continua respondendo `200`: negá-la inteira tiraria da
        chave os dados do contrato, que ela pode ver.

        Não confundir `bomcontrole_invoices` (o extrato) com `bomcontrole` (o
        **vínculo** do contrato com o ERP), que é outro campo e está sempre
        presente.
      operationId: getContract
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Consulta realizada.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/Contrato'
                          - type: object
                            properties:
                              domains:
                                type: array
                                items: { $ref: '#/components/schemas/DominioContratado' }
                              invoices:
                                type: array
                                description: Faturas geradas pelo CDW Finance. Só com o escopo `invoices:read`.
                                items: { $ref: '#/components/schemas/Fatura' }
                              bomcontrole_invoices:
                                allOf:
                                  - $ref: '#/components/schemas/ExtratoBomControle'
                                  - description: Extrato do ERP. Só com o escopo `invoices:read`.
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '403': { $ref: '#/components/responses/SemPermissao' }
        '404': { $ref: '#/components/responses/NaoEncontrado' }
        '422': { $ref: '#/components/responses/NaoProcessavel' }
        '429': { $ref: '#/components/responses/LimiteExcedido' }
...
  /servers/{id}:
    get:
      tags: [Infraestrutura]
      summary: Consulta um servidor por ID
      description: Requer o escopo `servers:read`. Servidor de outra empresa responde `404`.
      operationId: getServer
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Consulta realizada.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Servidor' }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '403': { $ref: '#/components/responses/SemPermissao' }
        '404': { $ref: '#/components/responses/NaoEncontrado' }
        '422': { $ref: '#/components/responses/NaoProcessavel' }
        '429': { $ref: '#/components/responses/LimiteExcedido' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Chave opaca gerada no painel (não é JWT), no formato
        `cdwf_live_<prefixo de 12 hex>_<segredo de 43 chars>`.
  schemas:
    Envelope:
      type: object
      description: Formato de toda resposta da API, com sucesso ou com erro.
      required: [success, message, data, errors]
      properties:
        success:
          type: boolean
          description: Indica se a operação foi concluída.
          example: true
        message:
          type: string
          description: Mensagem legível, em PT-BR.
          example: Operação realizada com sucesso.
        data:
          nullable: true
          description: Conteúdo da resposta. `null` quando não há retorno.
        errors:
          type: object
          additionalProperties:
            type: string
          description: Erros por campo. Objeto vazio quando não há erro.
          example: {}
        meta:
          $ref: '#/components/schemas/Meta'
    Meta:
      type: object
      description: Presente apenas em listagens paginadas.
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          example: 137
...
    Cliente:
      type: object
      description: |
        Cliente final do tenant. Os dados do representante legal (RG, CPF,
        endereço residencial), o contato financeiro do cadastro, o aceite de
        LGPD e a procedência interna **não são expostos**.
      properties:
        id: { type: integer, example: 1059 }
        type: { type: string, enum: [F, J] }
        type_label: { type: string, example: Pessoa jurídica }
        document: { type: string, description: CPF ou CNPJ, só dígitos. }
        name: { type: string }
        byname: { type: string, nullable: true, description: Nome fantasia. }
        state_registration: { type: string, description: 'Inscrição estadual; `ISENTO` quando não há.' }
        email: { type: string, nullable: true }
        address: { $ref: '#/components/schemas/Endereco' }
        declared_domains:
          type: object
          description: Domínios informados pelo cliente no cadastro — não é o cadastro de domínios do contrato.
          properties:
            primary: { type: string, nullable: true }
            secondary: { type: string, nullable: true }
        contract_notes: { type: string, nullable: true }
        contracts_count: { type: integer }
        active_contracts_count: { type: integer, description: Contratos com status `vigente`. }
        service_type_ids:
          type: array
          description: Ids dos tipos de serviço já contratados, de contratos de qualquer situação. Traduza em `/service-types`.
          items: { type: integer }
        created_at: { type: string, nullable: true, format: date-time }
        updated_at: { type: string, nullable: true, format: date-time }
...
  parameters:
    id:
      name: id
      in: path
      required: true
      description: Identificador do recurso.
      schema:
        type: integer
        minimum: 1
...
  responses:
    NaoAutenticado:
      description: Chave ausente, malformada, inválida ou revogada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
          example:
            success: false
            message: Chave de API inválida.
            data: null
            errors:
              authorization: Chave de API inválida.
    SemPermissao:
      description: A chave é válida, mas não possui o escopo exigido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
          example:
            success: false
            message: Esta chave não possui a permissão necessária.
            data: null
            errors:
              scope: 'Escopo obrigatório: invoices:read'
    NaoEncontrado:
      description: Recurso inexistente ou fora do escopo da empresa da chave.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
          example:
            success: false
            message: Endpoint não encontrado.
            data: null
            errors:
              path: Nenhum endpoint da API v1 corresponde a este caminho.
    NaoProcessavel:
      description: Requisição bem formada, mas com dados que não passam na validação.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
          example:
            success: false
            message: Parâmetro limit inválido.
            data: null
            errors:
              limit: Informe um inteiro entre 1 e 100.
    LimiteExcedido:
      description: Limite de requisições da chave excedido.
      headers:
        Retry-After:
          description: Segundos até a próxima tentativa permitida.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
          example:
            success: false
            message: Limite de requisições excedido.
            data: null
            errors:
              rate_limit: Tente novamente em 42 segundos.